Move discotron actions to endpoint.
